Punt to using strerror for this sample implementation.
GNU preferrs its own string returning version of strerror_r, which
might not even be safe, and is hard to replace with defines.  Once
upon a time, and maybe still, NGINX and other systems simply create
their own version by iterating over the raw array of error message
strings; this was considered to be the most sane way of doing it.